chmod +x helper scripts to avoid git's dirty flag
mk_c2kimage.sh and uldr_gen.sh are made executable (chmod +x) during the
build process which would cause git to flag them as modified. As a
result, git would describe the working directory as "dirty" and it
appends the "-dirty" suffix to the version string which is misleading.