Several changes to reduce binary size of uloader.
clkcore: don't inline things. Saves ~2k.

Allow MALLOC_DUMMY even if using a shell.  This is a bit ill-advised if you
plan to use the shell a *lot*, but it works okay for running a command or
three.

Allow disabling of environment variables evenn when the shell is enabled.

comcerto_nand: only include write functions if CONFIG_NAND_WRITE.

Change-Id: Ia99a0e3d3a9b89e6dea3529f3831f5445fea6f42