Mindspeed's patch to add a "-a" option to loady.
This allows the ymodem downloader to download straight to a memory location
instead of a file.  Useful in the microloader so it doesn't have to have a
filesystem.

Note: I modified this patch from the original, which required -a and -o to
be provided separately.  If you didn't provide -o, the default offset was 0,
which was pretty useless (and confusing since in "loady -a 0x01000000" the
second parameter was completely ignored).  Now the offset is mandatory so
it's harder to mess up.

Change-Id: I610ca731d806a51752ab711db4c98c16feab449d