Switch back to hardware ECC, but disable the badblock table.
We will probably want to switch back to the badblock table eventually,
because the hardware ECC seems to actually reuse the part of the OOB area
where the manufacturer bad block indicators are stored, which is kind of
silly. It then goes and puts the actual badblock indicators in a different
location, which is cheating.

Change-Id: I058422398bbd679d139650aba61b3a74eaa82aca