commands/hnvram: new command for reading vars from hnvram partitions.
This partition format is compatible with other GFiberTV devices.  The
current implementation of the command is read-only; it knows how to read
variables from the hnvram partition, and populates them into the barebox
environment.  We prepend HNV_ to all variable names in order to prevent
accidentally overwriting an "important" environment variable.

Change-Id: Ib5cf6ba115f6066cf421bb699529237c18b7ac5d