bootm: add minimal support for multi-file uImages.
This allows us to put the kernel and initrd into a single uImage, like you
can with u-boot.  There are patches that more or less go in this direction
in later versions of barebox, but they're too hard to backport.  We'll
probably end up throwing this away if we update barebox.

Change-Id: Ic87c8188bdd481e4c3e0f23bfa1ecfc425122484