Move stack and malloc area to 0x05000000
Depending on whether or not CONFIG_ZONE_DMA is set in the Linux kernel,
it copies itself to either 0x04008000 (CONFIG_ZONE_DMA set) or
0x00008000 (CONFIG_ZONE_DMA not set). Let's get out of the way and use
anything from 0x05000000 to 0x10000000 for our stack and malloc area.
Usually, it would not matter if Linux overwrote the barebox malloc area.
However, since we load the initrd into the malloc area, we need to make
sure it is not overwritten by kernel text, data etc.

Change-Id: I5fd3165e020135c8a98c102ae76ebb193c14499f