Enable barebox to authenticate kernel images.
If OTP auth is enabled (or CONFIG_FORCE_KERNEL_AUTH is set),
barebox will attempt to authenticate the kernel image against
the public key stored in include/rsa_public_key.h.

The change is backwards compatible, in that the barebox can
still boot the current kernel image format. Note that it is
not possible to authenticate current kernel images, so if
authentication is expected and the current ('legacy') kernel
format is detected, the boot will fail.

Change-Id: Ica6ece47d0b7bbaf7c047f2fba7bf95da36f0bbe