Decrease size of uloader stack to avoid collision
Mindspeed dedicated the IRAM area from 0x8300FC00 to 0x8300FD74 to store
the bypass bits for a variety of clocks like the AXI or A9DP clock. The
bypass bits determine whether the clock divider output is used or
bypassed. Due to a silicon bug these values can only be written but not
read back from the respective registers by software. As a workaround,
the uloader uses said memory area in IRAM to communicate these settings
to barebox and the Linux kernel.

Previously, this memory area overlapped with the area that is used for
the stack. This resulted in corruption of the bypass bits. As a result,
barebox made wrong assumptions about clock rates and configured
components incorrectly. For example, the UART divisor got misconfigured
because barebox assumed an AXI clock rate of 1 GHz instead of 250 MHz.

This commit reduces the size of the stack in the uloader from 16kB to
15kB to avoid this type of data corruption.

Change-Id: Ic9ae982569a23ff32bfd709dc9f5c1d67261ed42