uloader: Move malloc area back to IRAM
Shrink the uloader stack from 15kB to 4kB to make space for moving the
malloc area from DRAM back into IRAM. This allows us to execute the
uloader completely out of IRAM. We no longer depend on DRAM.

We previously needed a bigger stack for fancy features like a command
line interface in the uloader including the loady command that would
allow us to load bootloaders over YMODEM. Since we disabled all these
features, we no longer need the bigger stack and can go back to
Mindspeed's original design.

We are still not allowed to touch anything beyond 0x8300FC00 because
that is where the bypass bits are stored and we should not corrupt
those. We therefore lose 1kB and have only 11kB instead of 12kB
available for malloc.

Change-Id: I7d25c5b40ca8e72fd1da7d2a80db9aeac24301c7