Framework for multiple DRAM configurations
Currently, we only allow a single DDRC configuration that is baked into
the uloader binary. To support multiple boards with different DRAM
configurations, we created this framework which detects the type of
board we are running on and sets up the DDR controller accordingly. At
this point, we only support Optimus. In a subsequent commit, we will add
support for Sideswipe.

Also, print the name of the board we detected. If we detect on unknown
board, the following error message will be printed on the serial
console:

SoC: Mindspeed C2000
Board: Unknown
Bad DDRC configuration
 ### ERROR ### Please RESET the board ###

Change-Id: I03e86c6a7415fbadc2aaa7061ed12259d6a5cb81