UART output for low-level debugging
When the bootloader crashes very early during intialization, the lack of
console output makes debugging difficult. This putc() function can be
used to write a single character to the console. It requires the UART to
be initialized, though. At this point, we just leverage the fact that
the uloader has already done so. If we wanted to use this in the
uloader, then we would have to add an extra routine to initialize the
UART early in the uloader.

This is not compiled in by default. If you want to use this, you can add
"#define CONFIG_DEBUG_LL" to include/debug_ll.h and then "#include
<debug_ll.h>" from your .c file. Take a look at common/startup.c for an
example.

Change-Id: I15d9d5db70e71110ebfd86d435a9a67aced00acb