Automatically load bootloader over UART/YMODEM
The system config register tells us whether the uloader has been loaded
over the UART. If that is the case, then also load the bootloader over
the UART using the YMODEM protocol. On locked boxes, the signature of
the bootloader will still be verified.

Set CONFIG_CMD_BOOTB_UART_DOWNLOAD=y to enable this feature.

Change-Id: Idd9e6a13c692ee904a3673cab679fd241e4b1c35