Delete rootfs-prep ubi volume when tftp upgrading
/bin/update is a script used by developers to update the kernel and
rootfs from barebox. It is not used in the field.

When ginstall crashes midway through the update, it leaves a UBI volume
named rootfs-prep behind. Previously, /bin/update could not handle this
situation and eventually failed with "write: No space left on device".
Let's just delete this UBI volume before we upgrade.

Change-Id: Ib5de01fc6949bc9d55f9112a57c86a0e68c88592