uloader: Try backup bootloader if primary corrupt
We store two copies of the bootloader in the NOR flash. Previously, the
uloader would only attempt to load the primary copy in the loader0
partition. Let's change the uloader in a way such that it falls back to
the secondary copy in the loader1 partition if it can't verify the
primary copy. While we're at it, let's also make multiple attempts to
download the booloader over UART if we're booting from the same.

Change-Id: I75cf63c028d098b38e708240158bc14b64382287