Remove shell from uloader
We no longer provide a shell in the uloader so let's remove it. This
required some refactoring because the core logic of our uloader was
enabled by CONFIG_CMD_BOOTB which depends on a shell and therefore can
no longer be enabled. CONFIG_CMD_BOOTB turned out to be a misnomer
anyway because it does not implement a bootb command. Going forward,
arch/arm/lib/bootb.c is compiled in if CONFIG_COMCERTO_ULOADER is set.

Change-Id: I62c0dea49fdf1f215340649fdac8f3ac95c514b2