uloader: Turn blue LED off, red LED on
In order to comply with the LED policy, we need to turn the blue LED off
and the red LED on when we initialize the uloader. The downside of this
is that we can no longer tell if the uloader or bootloader crashed.
Previously, people confused solid blue with the state where the unit
successfully connected to the ACS.

Change-Id: I1d5eed20470436d81fa15cf45f241f498d0775a7