Move get_board_id to optimus.c and add optimus.h
We are introducing another variant of Optimus board
type, namely SpaceCast. The board initialization is
based on the board type. As a result, optimus.c needs
to use get_board_id. Therefore, we move this function
from ddr.c to optimus.c. This function is declared in
optimus.h. The ddr.c will include optimus.h to avoid
getting the compiler warning of implicit declaration
of function.

Change-Id: I683a978f97dd7fe237e38a80d3fdcbc5356fdf66