GPIO 48: Set direction first to avoid glitch
GPIO 48 controls power for the 11AC power applifier. Previously, we set
the direction of the GPIO pin to output and only subsequently set it to
high (the default is low). This could result in a glitch where the GPIO
pin goes from tri-state to low for a short period of time before it goes
to high and stays high. Avoid this glitch and make it go from tri-state
straight to high.

Change-Id: Idc1b06b307091adbcc82a799c22ff6b700f25905