Optimus Prime: GPIO setup for POE_ACTIVE,POE_RESET
On Optimus Prime and Sideswipe Prime, we have

POE_ACTIVE on GPIO 49: This is an input signal. It is low if a PoE
powered device is connected to the WAN port.

POE_RESET on GPIO 50: This is an output signal. It is pulled low on the
board. Software can drive it high if it wants to cut PoE power i.e. to
power cycle a PoE powered device (Fiber Jack).

Change-Id: I70fa5e220bb9ecd0d3866d37ad51dc8171f12cde